Рефакторит ошибки при загрузке фотографии, добавляет отображение загруженного пользователем изображения
