+++ js/main.js
@@ -4,5 +4,6 @@ import './form-validate';
 import './photo-scale';
 import './slider-effects';
 import './sort';
+import './photo-preview';
 
 getServerData(getPictures);
